+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Fixed
+- SETUP / TEARDOWN keyword removing, by @HardNorth
+
+## [5.6.1]
 ### Added
 - `RP_FLATTEN_KEYWORDS` configuration variable, by @HardNorth
 - `--flatten-keywords` argument support, by @HardNorth